countbb/requirements.txt
backcall==0.2.0
decorator==5.1.1
ghidra-bridge==0.2.5
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jfx-bridge==0.8.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.26
ptyprocess==0.7.0
Pygments==2.11.2
PyYAML==6.0
six==1.16.0
traitlets==4.3.3
wcwidth==0.2.5
